Move isOnPrem to core utils
