+++ public/app/core/utils/isOnPrem.ts
